refactor(WinningNumbers) 폴더 구조 변경
