+++ src/test/java/lotto/model/LottoAggregatorTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import org.junit.jupiter.api.Test;
