+++ src/test/java/lotto/model/LottosTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import lotto.model.LottoNumbers;
 import lotto.model.Lottos;
